+++ src/appa/odom.cpp
@@ -3,7 +3,7 @@
 namespace appa {
 
 /* Odom */
-Odom::Odom(int8_t x_port, int8_t y_port, int8_t imu_port, int tpi, Point tracker_linear_offset,
+Odom::Odom(int8_t x_port, int8_t y_port, int8_t imu_port, double tpi, Point tracker_linear_offset,
            double tracker_angular_offset)
     : x_tracker(abs(x_port), abs(x_port) + 1, x_port < 0),
       y_tracker(abs(y_port), abs(y_port) + 1, y_port < 0),
@@ -12,7 +12,7 @@ Odom::Odom(int8_t x_port, int8_t y_port, int8_t imu_port, int tpi, Point tracker
       tracker_linear_offset(tracker_linear_offset),
       tracker_angular_offset(to_rad(tracker_angular_offset)) {}
 
-Odom::Odom(std::array<int8_t, 2> x_port, std::array<int8_t, 2> y_port, int8_t imu_port, int tpi,
+Odom::Odom(std::array<int8_t, 2> x_port, std::array<int8_t, 2> y_port, int8_t imu_port, double tpi,
            Point tracker_linear_offset, double tracker_angular_offset)
     : x_tracker({x_port[0], abs(x_port[1]), abs(x_port[1]) + 1}, x_port[1] < 0),
       y_tracker({y_port[0], abs(y_port[1]), abs(y_port[1]) + 1}, y_port[1] < 0),
